fix: service import and throws exception

+++ src/main/java/app/trybe/specialityapp/service/ProfessionalService.java
@@ -6,8 +6,8 @@ import app.trybe.specialityapp.repository.ProfessionalRepository;
 import java.util.List;
 import java.util.Optional;
 import javax.ws.rs.core.Response;
-import org.jvnet.hk2.annotations.Service;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 @Service
 public class ProfessionalService implements ServiceInterface<Professional, Integer> {
@@ -15,7 +15,7 @@ public class ProfessionalService implements ServiceInterface<Professional, Integ
   private ProfessionalRepository professionalRepository;
 
   @Override
-  public String save(Professional p) {
+  public String save(Professional p) throws ApplicationError {
     if (p.getId() != null) {
       throw new ApplicationError(Response.Status.BAD_REQUEST,
               "Não é permitido inserir novos registros com ID explícito");
@@ -27,7 +27,7 @@ public class ProfessionalService implements ServiceInterface<Professional, Integ
   }
 
   @Override
-  public String update(Professional p, int id) {
+  public String update(Professional p, int id) throws ApplicationError {
     Optional<Professional> professional = professionalRepository.findById(id);
 
     if (professional.isPresent()) {
@@ -41,19 +41,21 @@ public class ProfessionalService implements ServiceInterface<Professional, Integ
   }
 
   @Override
-  public String delete(int id) {
-    try {
+  public String delete(int id) throws ApplicationError {
+    Optional<Professional> professional = professionalRepository.findById(id);
+
+    if (professional.isPresent()) {
       professionalRepository.deleteById(id);
 
-      return String.format("ID [%d] atualizado", id);
-    } catch (Error e) {
-      throw new ApplicationError(Response.Status.NOT_FOUND,
-              "Não é possível deletar, o ID informado não existe");
+      return String.format("ID [%d] removido", id);
     }
+
+    throw new ApplicationError(Response.Status.NOT_FOUND,
+            "Não é possível deletar, o ID informado não existe");
   }
 
   @Override
-  public List<Professional> list() {
+  public List<Professional> list() throws ApplicationError {
     List<Professional> profs = professionalRepository.findAll();
 
     if (profs == null || profs.isEmpty()) {
